% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mhde.R
\name{mhde.cn}
\alias{mhde.cn}
\title{Utility function for retrieving \eqn{c_n}{Cn} value for any sample size}
\usage{
mhde.cn(Nsize)
}
\arguments{
\item{Nsize}{Number of values in the sample data set}
}
\value{
The Epanechnikov kernel bandwidth parameter for the sample size.
}
\description{
This function evaluates the \eqn{c_n}{Cn} bandwidth value for the Epanechnikov kernel based on the sample size.
}

