% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mhde.R
\name{mhde.plot}
\alias{mhde.plot}
\title{Plot the non-parametric and normal densities}
\usage{
mhde.plot(ListOut)
}
\arguments{
\item{ListOut}{The output list from \code{mhde.test}.  The argument \code{Small} for \code{mhde.test} must have the value FALSE.}
}
\description{
Simple plot function to display the data-based kernel density and the normal density for the final location and scale estimates.
}

